reorganize files and add docstring, edit readme file
