+++ dockerfile
@@ -2,7 +2,7 @@ FROM python:3.11
 
 RUN mkdir /src
 
-COPY * /src
+COPY ./src /src
 
 WORKDIR /src
 
@@ -10,6 +10,6 @@ ENV PYTHONPATH=${PYTHONPATH}:${PWD}
 
 RUN pip3 install poetry
 RUN poetry config virtualenvs.create false
-RUN poetry install
+RUN poetry install --no-dev
 
 CMD ["./docker-entrypoint.sh"]
